requirements.txt
africastalking==1.2.8
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-oauth-toolkit==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.9
jwcrypto==1.5.6
jwt==1.3.1
oauthlib==3.2.2
phonenumbers==8.13.45
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
schema==0.7.7
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
